chore: Update type file definition and remove unused parameter in middleware
