+++ components/common/AvatarIcon.tsx
@@ -1,6 +1,6 @@
 import { Skeleton } from "@/components/ui/skeleton";
 import { useUser } from "@/hooks/useUser";
-import type { User } from "@/lib/types";
+import type { User } from "@/lib/types/user";
 import { cn } from "@/lib/utils";
 import { md5 } from "@/lib/utils/md5";
 import { Avatar, AvatarFallback, AvatarImage } from "@radix-ui/react-avatar";
