+++ components/home/CreateBoardForm.tsx
@@ -4,7 +4,7 @@ import { BoardState } from "@/db/schema";
 import { authenticatedCreateBoard } from "@/lib/actions/authenticatedActions";
 import { addBoard, removeBoard } from "@/lib/signal/boardSignals";
 import { toast } from "@/lib/signal/toastSignals";
-import type { NewBoard } from "@/lib/types";
+import type { NewBoard } from "@/lib/types/board";
 import { nanoid } from "nanoid";
 import type { FormEvent } from "react";
