+++ lib/types/action.ts
@@ -1,0 +1,4 @@
+import type { actionsTable } from "@/db/schema";
+
+export type NewAction = typeof actionsTable.$inferInsert;
+export type Action = typeof actionsTable.$inferSelect;
