+++ lib/types/board.ts
@@ -1,0 +1,4 @@
+import type { boardTable } from "@/db/schema";
+
+export type NewBoard = typeof boardTable.$inferInsert;
+export type Board = typeof boardTable.$inferSelect;
