+++ lib/types/user.ts
@@ -1,0 +1,4 @@
+import type { userTable } from "@/db/schema";
+
+export type NewUser = typeof userTable.$inferInsert;
+export type User = typeof userTable.$inferSelect;
